auth/requirements.txt
flask==2.2.5
psycopg2-binary==2.9.7
werkzeug==2.2.3
gunicorn==20.1.0
flask-cors==3.0.10
